requirements.txt
click==8.0.1
mypy-extensions==0.4.3
numpy==1.21.2
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.4.0
regex==2021.9.24
tomli==1.2.1
typing-extensions==3.10.0.2
